Refactor the code to use session and module abstractions for build system and test framework
