+++ dd-java-agent/agent-ci-visibility/build.gradle
@@ -12,12 +12,14 @@ minimumInstructionCoverage = 0.8
 excludedClassesCoverage += [
   "datadog.trace.civisibility.CiVisibilitySystem",
   "datadog.trace.civisibility.CiVisibilitySystem.1",
-  "datadog.trace.civisibility.DDTestModuleChild",
+  "datadog.trace.civisibility.DDBuildSystemModuleImpl",
+  "datadog.trace.civisibility.DDBuildSystemSessionImpl",
+  "datadog.trace.civisibility.DDTestFrameworkModuleImpl",
+  "datadog.trace.civisibility.DDTestFrameworkModuleProxy",
+  "datadog.trace.civisibility.DDTestFrameworkSessionImpl",
+  "datadog.trace.civisibility.DDTestFrameworkSessionProxy",
   "datadog.trace.civisibility.DDTestModuleImpl",
-  "datadog.trace.civisibility.DDTestModuleParent",
-  "datadog.trace.civisibility.DDTestSessionChild",
   "datadog.trace.civisibility.DDTestSessionImpl",
-  "datadog.trace.civisibility.DDTestSessionParent",
   "datadog.trace.civisibility.DDTestSuiteImpl",
   "datadog.trace.civisibility.DDTestImpl",
   "datadog.trace.civisibility.TestModuleRegistry",
@@ -65,6 +67,7 @@ excludedClassesCoverage += [
   "datadog.trace.civisibility.ipc.SignalServer",
   "datadog.trace.civisibility.ipc.SignalType",
   "datadog.trace.civisibility.ipc.SignalClient",
+  "datadog.trace.civisibility.ipc.SignalClient.Factory",
   "datadog.trace.civisibility.ipc.SkippableTestsResponse",
   "datadog.trace.civisibility.source.MethodLinesResolver.MethodLines",
   "datadog.trace.civisibility.source.index.RepoIndexBuilder.RepoIndexingFileVisitor",
