+++ dd-java-agent/agent-ci-visibility/src/main/java/datadog/trace/civisibility/DDBuildSystemSession.java
@@ -1,5 +1,6 @@
 package datadog.trace.civisibility;
 
+import datadog.trace.api.civisibility.DDTestSession;
 import datadog.trace.api.civisibility.config.ModuleExecutionSettings;
 import datadog.trace.civisibility.config.JvmInfo;
 import java.io.File;
@@ -7,12 +8,8 @@ import java.nio.file.Path;
 import java.util.Collection;
 import javax.annotation.Nullable;
 
-public interface DDBuildSystemSession {
-  void setTag(String key, Object value);
-
-  void setErrorInfo(Throwable error);
-
-  void end(@Nullable Long endTime);
+/** Test session abstraction that is used by build system instrumentations (e.g. Maven, Gradle) */
+public interface DDBuildSystemSession extends DDTestSession {
 
   DDBuildSystemModule testModuleStart(
       String moduleName, @Nullable Long startTime, Collection<File> outputClassesDirs);
@@ -21,6 +18,10 @@ public interface DDBuildSystemSession {
 
   interface Factory {
     DDBuildSystemSession startSession(
-        String projectName, Path projectRoot, String component, Long startTime);
+        String projectName,
+        Path projectRoot,
+        String startCommand,
+        String buildSystemName,
+        Long startTime);
   }
 }
