+++ dd-java-agent/agent-ci-visibility/src/main/java/datadog/trace/civisibility/DDTestFrameworkSession.java
@@ -1,5 +1,16 @@
 package datadog.trace.civisibility;
 
+import java.nio.file.Path;
+import javax.annotation.Nullable;
+
+/** Test session abstraction that is used by test framework instrumentations (e.g. JUnit, TestNG) */
 public interface DDTestFrameworkSession {
   void end(Long startTime);
+
+  DDTestFrameworkModule testModuleStart(String moduleName, @Nullable Long startTime);
+
+  interface Factory {
+    DDTestFrameworkSession startSession(
+        String projectName, Path projectRoot, String component, Long startTime);
+  }
 }
