+++ dd-java-agent/agent-ci-visibility/src/main/java/datadog/trace/civisibility/DDTestFrameworkSessionProxy.java
@@ -1,20 +1,21 @@
 package datadog.trace.civisibility;
 
 import datadog.trace.api.Config;
-import datadog.trace.api.civisibility.config.ModuleExecutionSettings;
 import datadog.trace.api.civisibility.coverage.CoverageDataSupplier;
 import datadog.trace.civisibility.codeowners.Codeowners;
-import datadog.trace.civisibility.config.JvmInfo;
 import datadog.trace.civisibility.coverage.CoverageProbeStoreFactory;
 import datadog.trace.civisibility.decorator.TestDecorator;
 import datadog.trace.civisibility.source.MethodLinesResolver;
 import datadog.trace.civisibility.source.SourcePathResolver;
-import java.io.File;
 import java.net.InetSocketAddress;
-import java.util.Collection;
 import javax.annotation.Nullable;
 
-public class DDTestSessionChild extends DDTestSessionImpl {
+/**
+ * Test session implementation that is used by test framework instrumentations in those cases when
+ * the build system IS instrumented: since build system instrumentation manages the session span,
+ * this class does not do it
+ */
+public class DDTestFrameworkSessionProxy implements DDTestFrameworkSession {
 
   private final Long parentProcessSessionId;
   private final Long parentProcessModuleId;
@@ -27,7 +28,7 @@ public class DDTestSessionChild extends DDTestSessionImpl {
   private final CoverageDataSupplier coverageDataSupplier;
   @Nullable private final InetSocketAddress signalServerAddress;
 
-  public DDTestSessionChild(
+  public DDTestFrameworkSessionProxy(
       Long parentProcessSessionId,
       Long parentProcessModuleId,
       Config config,
@@ -51,29 +52,13 @@ public class DDTestSessionChild extends DDTestSessionImpl {
   }
 
   @Override
-  public void setTag(String key, Object value) {
-    throw new UnsupportedOperationException("Setting tags is not supported: " + key + ", " + value);
-  }
-
-  @Override
-  public void setErrorInfo(Throwable error) {
-    throw new UnsupportedOperationException("Setting error info is not supported: " + error);
-  }
-
-  @Override
-  public void setSkipReason(String skipReason) {
-    throw new UnsupportedOperationException("Setting skip reason is not supported: " + skipReason);
-  }
-
-  @Override
-  public void end(@Nullable Long endTime) {
+  public void end(Long startTime) {
     // no op
   }
 
   @Override
-  public DDTestModuleImpl testModuleStart(
-      String moduleName, @Nullable Long startTime, Collection<File> outputClassesDirs) {
-    return new DDTestModuleChild(
+  public DDTestFrameworkModule testModuleStart(String moduleName, @Nullable Long startTime) {
+    return new DDTestFrameworkModuleProxy(
         parentProcessSessionId,
         parentProcessModuleId,
         moduleName,
@@ -85,11 +70,5 @@ public class DDTestSessionChild extends DDTestSessionImpl {
         coverageProbeStoreFactory,
         coverageDataSupplier,
         signalServerAddress);
-  }
-
-  @Override
-  public ModuleExecutionSettings getModuleExecutionSettings(JvmInfo jvmInfo) {
-    throw new UnsupportedOperationException(
-        "Getting module execution settings is not supported: " + jvmInfo);
   }
 }
