+++ dd-java-agent/agent-ci-visibility/src/main/java/datadog/trace/civisibility/source/index/RepoIndexFetcher.java
@@ -3,7 +3,6 @@ package datadog.trace.civisibility.source.index;
 import datadog.trace.civisibility.ipc.RepoIndexRequest;
 import datadog.trace.civisibility.ipc.RepoIndexResponse;
 import datadog.trace.civisibility.ipc.SignalClient;
-import java.util.function.Supplier;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -11,11 +10,11 @@ public class RepoIndexFetcher implements RepoIndexProvider {
 
   private static final Logger log = LoggerFactory.getLogger(RepoIndexFetcher.class);
 
-  private final Supplier<SignalClient> signalClientFactory;
+  private final SignalClient.Factory signalClientFactory;
   private final Object indexInitializationLock = new Object();
   private volatile RepoIndex index;
 
-  public RepoIndexFetcher(Supplier<SignalClient> signalClientFactory) {
+  public RepoIndexFetcher(SignalClient.Factory signalClientFactory) {
     this.signalClientFactory = signalClientFactory;
   }
 
@@ -32,7 +31,7 @@ public class RepoIndexFetcher implements RepoIndexProvider {
   }
 
   private RepoIndex doGetIndex() {
-    try (SignalClient signalClient = signalClientFactory.get()) {
+    try (SignalClient signalClient = signalClientFactory.create()) {
       RepoIndexResponse response = (RepoIndexResponse) signalClient.send(RepoIndexRequest.INSTANCE);
       return response.getIndex();
     } catch (Exception e) {
