+++ dd-trace-api/build.gradle
@@ -18,9 +18,6 @@ excludedClassesCoverage += [
   'datadog.trace.api.internal.TraceSegment',
   'datadog.trace.api.internal.TraceSegment.NoOp',
   'datadog.trace.api.civisibility.CIVisibility',
-  'datadog.trace.api.civisibility.config.Configurations',
-  'datadog.trace.api.civisibility.DDTestModule',
-  'datadog.trace.api.civisibility.config.SkippableTest',
   'datadog.trace.api.config.ProfilingConfig',
   'datadog.trace.api.interceptor.MutableSpan',
   'datadog.trace.api.experimental.Profiling',
