+++ dd-trace-api/src/main/java/datadog/trace/api/civisibility/DDTestModule.java
@@ -1,6 +1,5 @@
 package datadog.trace.api.civisibility;
 
-import datadog.trace.api.civisibility.config.SkippableTest;
 import javax.annotation.Nullable;
 
 /**
@@ -89,12 +88,4 @@ public interface DDTestModule {
       @Nullable Class<?> testClass,
       @Nullable Long startTime,
       boolean parallelized);
-
-  /**
-   * Checks if a given test can be skipped with Intelligent Test Runner or not
-   *
-   * @param test Test to be checked
-   * @return {@code true} if the test can be skipped, {@code false} otherwise
-   */
-  boolean skip(SkippableTest test);
 }
